REF: move Series arithmetic helpers
